requirements.txt
Flask==3.0.0
gunicorn==20.1.0
tensorflow==2.8.0
numpy==1.22.3
Pillow==9.0.1
matplotlib==3.5.1
seaborn==0.11.2
Werkzeug==3.0.0

